+++ lib/src/models/mastodon_collection.freezed.dart
@@ -15,7 +15,7 @@ T _$identity<T>(T value) => value;
 /// @nodoc
 mixin _$MastodonCollection {
 
- String get id; String? get uri; String get name; String? get description; String? get language; String? get accountId; bool? get local; bool? get sensitive; bool? get discoverable; String? get url; int? get itemCount; DateTime? get createdAt; DateTime? get updatedAt; String? get tag;
+ String get id; String? get uri; String get name; String? get description; String? get language; String? get accountId; bool? get local; bool? get sensitive; bool? get discoverable; String? get url; int? get itemCount; DateTime? get createdAt; DateTime? get updatedAt; MastodonCollectionTag? get tag; List<MastodonCollectionItem> get items;
 /// Create a copy of MastodonCollection
 /// with the given fields replaced by the non-null parameter values.
 @JsonKey(includeFromJson: false, includeToJson: false)
@@ -26,12 +26,12 @@ $MastodonCollectionCopyWith<MastodonCollection> get copyWith => _$MastodonCollec
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is MastodonCollection&&(identical(other.id, id) || other.id == id)&&(identical(other.uri, uri) || other.uri == uri)&&(identical(other.name, name) || other.name == name)&&(identical(other.description, description) || other.description == description)&&(identical(other.language, language) || other.language == language)&&(identical(other.accountId, accountId) || other.accountId == accountId)&&(identical(other.local, local) || other.local == local)&&(identical(other.sensitive, sensitive) || other.sensitive == sensitive)&&(identical(other.discoverable, discoverable) || other.discoverable == discoverable)&&(identical(other.url, url) || other.url == url)&&(identical(other.itemCount, itemCount) || other.itemCount == itemCount)&&(identical(other.createdAt, createdAt) || other.createdAt == createdAt)&&(identical(other.updatedAt, updatedAt) || other.updatedAt == updatedAt)&&(identical(other.tag, tag) || other.tag == tag));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is MastodonCollection&&(identical(other.id, id) || other.id == id)&&(identical(other.uri, uri) || other.uri == uri)&&(identical(other.name, name) || other.name == name)&&(identical(other.description, description) || other.description == description)&&(identical(other.language, language) || other.language == language)&&(identical(other.accountId, accountId) || other.accountId == accountId)&&(identical(other.local, local) || other.local == local)&&(identical(other.sensitive, sensitive) || other.sensitive == sensitive)&&(identical(other.discoverable, discoverable) || other.discoverable == discoverable)&&(identical(other.url, url) || other.url == url)&&(identical(other.itemCount, itemCount) || other.itemCount == itemCount)&&(identical(other.createdAt, createdAt) || other.createdAt == createdAt)&&(identical(other.updatedAt, updatedAt) || other.updatedAt == updatedAt)&&(identical(other.tag, tag) || other.tag == tag)&&const DeepCollectionEquality().equals(other.items, items));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hash(runtimeType,id,uri,name,description,language,accountId,local,sensitive,discoverable,url,itemCount,createdAt,updatedAt,tag);
+int get hashCode => Object.hash(runtimeType,id,uri,name,description,language,accountId,local,sensitive,discoverable,url,itemCount,createdAt,updatedAt,tag,const DeepCollectionEquality().hash(items));
 
 
 
@@ -42,7 +42,7 @@ abstract mixin class $MastodonCollectionCopyWith<$Res>  {
   factory $MastodonCollectionCopyWith(MastodonCollection value, $Res Function(MastodonCollection) _then) = _$MastodonCollectionCopyWithImpl;
 @useResult
 $Res call({
- String id, String? uri, String name, String? description, String? language, String? accountId, bool? local, bool? sensitive, bool? discoverable, String? url, int? itemCount, DateTime? createdAt, DateTime? updatedAt, String? tag
+ String id, String? uri, String name, String? description, String? language, String? accountId, bool? local, bool? sensitive, bool? discoverable, String? url, int? itemCount, DateTime? createdAt, DateTime? updatedAt, MastodonCollectionTag? tag, List<MastodonCollectionItem> items
 });
 
 
@@ -59,7 +59,7 @@ class _$MastodonCollectionCopyWithImpl<$Res>
 
 /// Create a copy of MastodonCollection
 /// with the given fields replaced by the non-null parameter values.
-@pragma('vm:prefer-inline') @override $Res call({Object? id = null,Object? uri = freezed,Object? name = null,Object? description = freezed,Object? language = freezed,Object? accountId = freezed,Object? local = freezed,Object? sensitive = freezed,Object? discoverable = freezed,Object? url = freezed,Object? itemCount = freezed,Object? createdAt = freezed,Object? updatedAt = freezed,Object? tag = freezed,}) {
+@pragma('vm:prefer-inline') @override $Res call({Object? id = null,Object? uri = freezed,Object? name = null,Object? description = freezed,Object? language = freezed,Object? accountId = freezed,Object? local = freezed,Object? sensitive = freezed,Object? discoverable = freezed,Object? url = freezed,Object? itemCount = freezed,Object? createdAt = freezed,Object? updatedAt = freezed,Object? tag = freezed,Object? items = null,}) {
   return _then(MastodonCollection(
 id: null == id ? _self.id : id // ignore: cast_nullable_to_non_nullable
 as String,uri: freezed == uri ? _self.uri : uri // ignore: cast_nullable_to_non_nullable
@@ -75,7 +75,8 @@ as String?,itemCount: freezed == itemCount ? _self.itemCount : itemCount // igno
 as int?,createdAt: freezed == createdAt ? _self.createdAt : createdAt // ignore: cast_nullable_to_non_nullable
 as DateTime?,updatedAt: freezed == updatedAt ? _self.updatedAt : updatedAt // ignore: cast_nullable_to_non_nullable
 as DateTime?,tag: freezed == tag ? _self.tag : tag // ignore: cast_nullable_to_non_nullable
-as String?,
+as MastodonCollectionTag?,items: null == items ? _self.items : items // ignore: cast_nullable_to_non_nullable
+as List<MastodonCollectionItem>,
   ));
 }
 
@@ -84,6 +85,374 @@ as String?,
 
 /// Adds pattern-matching-related methods to [MastodonCollection].
 extension MastodonCollectionPatterns on MastodonCollection {
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
+
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>({required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _:
+  return orElse();
+
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
+
+@optionalTypeArgs TResult map<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
+
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
+
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  return null;
+
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
+
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>({required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _:
+  return orElse();
+
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
+
+@optionalTypeArgs TResult when<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
+
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
+
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  return null;
+
+}
+}
+
+}
+
+
+/// @nodoc
+mixin _$MastodonCollectionTag {
+
+ String get name; String get url;
+/// Create a copy of MastodonCollectionTag
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$MastodonCollectionTagCopyWith<MastodonCollectionTag> get copyWith => _$MastodonCollectionTagCopyWithImpl<MastodonCollectionTag>(this as MastodonCollectionTag, _$identity);
+
+
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is MastodonCollectionTag&&(identical(other.name, name) || other.name == name)&&(identical(other.url, url) || other.url == url));
+}
+
+@JsonKey(includeFromJson: false, includeToJson: false)
+@override
+int get hashCode => Object.hash(runtimeType,name,url);
+
+
+
+}
+
+/// @nodoc
+abstract mixin class $MastodonCollectionTagCopyWith<$Res>  {
+  factory $MastodonCollectionTagCopyWith(MastodonCollectionTag value, $Res Function(MastodonCollectionTag) _then) = _$MastodonCollectionTagCopyWithImpl;
+@useResult
+$Res call({
+ String name, String url
+});
+
+
+
+
+}
+/// @nodoc
+class _$MastodonCollectionTagCopyWithImpl<$Res>
+    implements $MastodonCollectionTagCopyWith<$Res> {
+  _$MastodonCollectionTagCopyWithImpl(this._self, this._then);
+
+  final MastodonCollectionTag _self;
+  final $Res Function(MastodonCollectionTag) _then;
+
+/// Create a copy of MastodonCollectionTag
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? name = null,Object? url = null,}) {
+  return _then(MastodonCollectionTag(
+name: null == name ? _self.name : name // ignore: cast_nullable_to_non_nullable
+as String,url: null == url ? _self.url : url // ignore: cast_nullable_to_non_nullable
+as String,
+  ));
+}
+
+}
+
+
+/// Adds pattern-matching-related methods to [MastodonCollectionTag].
+extension MastodonCollectionTagPatterns on MastodonCollectionTag {
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
+
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>({required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _:
+  return orElse();
+
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
+
+@optionalTypeArgs TResult map<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
+
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
+
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(){
+final _that = this;
+switch (_that) {
+case _:
+  return null;
+
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
+
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>({required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _:
+  return orElse();
+
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
+
+@optionalTypeArgs TResult when<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  throw StateError('Unexpected subclass');
+
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
+
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>() {final _that = this;
+switch (_that) {
+case _:
+  return null;
+
+}
+}
+
+}
+
+
+/// @nodoc
+mixin _$MastodonCollectionItem {
+
+ String get id; String get state; DateTime get createdAt; String? get accountId;
+/// Create a copy of MastodonCollectionItem
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$MastodonCollectionItemCopyWith<MastodonCollectionItem> get copyWith => _$MastodonCollectionItemCopyWithImpl<MastodonCollectionItem>(this as MastodonCollectionItem, _$identity);
+
+
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is MastodonCollectionItem&&(identical(other.id, id) || other.id == id)&&(identical(other.state, state) || other.state == state)&&(identical(other.createdAt, createdAt) || other.createdAt == createdAt)&&(identical(other.accountId, accountId) || other.accountId == accountId));
+}
+
+@JsonKey(includeFromJson: false, includeToJson: false)
+@override
+int get hashCode => Object.hash(runtimeType,id,state,createdAt,accountId);
+
+
+
+}
+
+/// @nodoc
+abstract mixin class $MastodonCollectionItemCopyWith<$Res>  {
+  factory $MastodonCollectionItemCopyWith(MastodonCollectionItem value, $Res Function(MastodonCollectionItem) _then) = _$MastodonCollectionItemCopyWithImpl;
+@useResult
+$Res call({
+ String id, String state, DateTime createdAt, String? accountId
+});
+
+
+
+
+}
+/// @nodoc
+class _$MastodonCollectionItemCopyWithImpl<$Res>
+    implements $MastodonCollectionItemCopyWith<$Res> {
+  _$MastodonCollectionItemCopyWithImpl(this._self, this._then);
+
+  final MastodonCollectionItem _self;
+  final $Res Function(MastodonCollectionItem) _then;
+
+/// Create a copy of MastodonCollectionItem
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? id = null,Object? state = null,Object? createdAt = null,Object? accountId = freezed,}) {
+  return _then(MastodonCollectionItem(
+id: null == id ? _self.id : id // ignore: cast_nullable_to_non_nullable
+as String,state: null == state ? _self.state : state // ignore: cast_nullable_to_non_nullable
+as String,createdAt: null == createdAt ? _self.createdAt : createdAt // ignore: cast_nullable_to_non_nullable
+as DateTime,accountId: freezed == accountId ? _self.accountId : accountId // ignore: cast_nullable_to_non_nullable
+as String?,
+  ));
+}
+
+}
+
+
+/// Adds pattern-matching-related methods to [MastodonCollectionItem].
+extension MastodonCollectionItemPatterns on MastodonCollectionItem {
 /// A variant of `map` that fallback to returning `orElse`.
 ///
 /// It is equivalent to doing: